UI: Merge VideBuffer into MainWindow/MainViewModel
