+++ src/Spice86/App.axaml
@@ -2,7 +2,9 @@
     x:Class="Spice86.App"
     xmlns="https://github.com/avaloniaui"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:Spice86">
+    RequestedThemeVariant="Default">
+    <!-- "Default" ThemeVariant follows system theme variant. "Dark" or "Light" are other available options. -->
+
     <Application.Styles>
         <FluentTheme />
     </Application.Styles>
