+++ src/Spice86/ViewModels/MainWindowViewModel.cs
@@ -32,6 +32,8 @@ using Spice86.Shared.Interfaces;
 using Key = Spice86.Shared.Emulator.Keyboard.Key;
 using MouseButton = Spice86.Shared.Emulator.Mouse.MouseButton;
 using Avalonia.Input.Platform;
+using Avalonia.Media.Imaging;
+using Avalonia.Platform;
 
 using Spice86.Shared.Diagnostics;
 
@@ -48,26 +50,105 @@ public sealed partial class MainWindowViewModel : ViewModelBase, IGui, IDisposab
     private PaletteWindow? _paletteWindow;
     private PerformanceWindow? _performanceWindow;
     private string _lastExecutableDirectory = string.Empty;
-
     private bool _closeAppOnEmulatorExit;
+
+    private Action? _uiUpdateMethod;
+    private bool _exitDrawThread;
+    private Action? _drawAction;
+    private Thread? _drawThread;
 
     public bool PauseEmulatorOnStart { get; private set; }
 
-    internal void OnKeyUp(KeyEventArgs e) => KeyUp?.Invoke(this, 
-        new((Key) e.Key, 
+    internal void OnKeyUp(KeyEventArgs e) => KeyUp?.Invoke(this,
+        new KeyboardEventArgs((Key) e.Key,
             false,
             _avaloniaKeyScanCodeConverter.GetKeyReleasedScancode((Key)e.Key),
             _avaloniaKeyScanCodeConverter.GetAsciiCode(_avaloniaKeyScanCodeConverter.GetKeyReleasedScancode((Key)e.Key))));
 
     private ProgramExecutor? _programExecutor;
 
-    [ObservableProperty]
-    private IVideoBufferViewModel? _videoBuffer;
+    [RelayCommand]
+    public async Task SaveBitmap() {
+        if (Application.Current?.ApplicationLifetime is IClassicDesktopStyleApplicationLifetime desktop &&
+            desktop.MainWindow is not null &&
+            desktop.MainWindow.StorageProvider.CanSave &&
+            desktop.MainWindow.StorageProvider.CanPickFolder) {
+            IStorageProvider storageProvider = desktop.MainWindow.StorageProvider;
+            FilePickerSaveOptions options = new() {
+                Title = "Save bitmap image...",
+                DefaultExtension = "bmp",
+                SuggestedStartLocation = await storageProvider.TryGetWellKnownFolderAsync(WellKnownFolder.Documents)
+            };
+            string? file = (await storageProvider.SaveFilePickerAsync(options))?.TryGetLocalPath();
+            if (!string.IsNullOrWhiteSpace(file)) {
+                Bitmap?.Save(file);
+            }
+        }
+    }
+
+    private bool _showCursor = false;
+
+    public bool ShowCursor {
+        get => _showCursor;
+        set {
+            SetProperty(ref _showCursor, value);
+            if (_showCursor) {
+                Cursor?.Dispose();
+                Cursor = Cursor.Default;
+            } else {
+                Cursor?.Dispose();
+                Cursor = new Cursor(StandardCursorType.None);
+            }
+        }
+    }
+
+    private double _scale = 1;
+
+    public double Scale {
+        get => _scale;
+        set => SetProperty(ref _scale, Math.Max(value, 1));
+    }
     
+    private void DrawThreadMethod() {
+        while (!_exitDrawThread) {
+            _drawAction?.Invoke();
+        }
+    }
+
+    /// <inheritdoc/>
+    public void Draw() {
+        if (_disposed || _isSettingResolution || _isAppClosing || _uiUpdateMethod is null || Bitmap is null || _videoCard is null) {
+            return;
+        }
+        if (_drawThread is null) {
+            _drawThread = new Thread(DrawThreadMethod) {
+                Name = "UIRenderThread"
+            };
+            _drawThread.Start();
+        }
+
+        _drawAction ??= () => {
+            unsafe {
+                using ILockedFramebuffer pixels = Bitmap.Lock();
+                var buffer = new Span<uint>((void*)pixels.Address, pixels.RowBytes * pixels.Size.Height / 4);
+                _videoCard.Render(buffer);
+                Dispatcher.UIThread.Post(() => _uiUpdateMethod.Invoke(), DispatcherPriority.Render);
+            }
+        };
+    }
+
+    public void SetUIUpdateMethod(Action uiUpdateMethod) => _uiUpdateMethod = uiUpdateMethod;
+
+    [ObservableProperty]
+    private Cursor? _cursor = Cursor.Default;
+
+    [ObservableProperty]
+    private WriteableBitmap? _bitmap;
+
     private ManualResetEvent _okayToContinueEvent = new(true);
 
-    internal void OnKeyDown(KeyEventArgs e) => KeyDown?.Invoke(this, 
-        new((Key) e.Key, 
+    internal void OnKeyDown(KeyEventArgs e) => KeyDown?.Invoke(this,
+        new KeyboardEventArgs((Key) e.Key,
             true,
             _avaloniaKeyScanCodeConverter.GetKeyPressedScancode((Key)e.Key),
             _avaloniaKeyScanCodeConverter.GetAsciiCode(_avaloniaKeyScanCodeConverter.GetKeyPressedScancode((Key)e.Key))));
@@ -90,13 +171,13 @@ public sealed partial class MainWindowViewModel : ViewModelBase, IGui, IDisposab
     public event EventHandler<MouseButtonEventArgs>? MouseButtonDown;
     public event EventHandler<MouseButtonEventArgs>? MouseButtonUp;
     
-    private bool _isMainWindowClosing;
+    private bool _isAppClosing;
 
     public MainWindowViewModel(Configuration configuration, ILoggerService loggerService) {
         Configuration = configuration;
         _loggerService = loggerService;
         if (Application.Current?.ApplicationLifetime is IClassicDesktopStyleApplicationLifetime desktop && desktop.MainWindow is not null) {
-            desktop.MainWindow.Closing += (_, _) => _isMainWindowClosing = true;
+            desktop.MainWindow.Closing += (_, _) => _isAppClosing = true;
         }
     }
 
@@ -105,19 +186,13 @@ public sealed partial class MainWindowViewModel : ViewModelBase, IGui, IDisposab
         PauseEmulatorOnStart = false;
     }
 
-    public void HideMouseCursor() => Dispatcher.UIThread.Post(() => {
-        if (VideoBuffer is null) {
-            return;
-        }
-        VideoBuffer.ShowCursor = false;
-    });
+    public int Width { get; private set; }
 
-    public void ShowMouseCursor() => Dispatcher.UIThread.Post(() => {
-        if (VideoBuffer is null) {
-            return;
-        }
-        VideoBuffer.ShowCursor = true;
-    });
+    public int Height { get; private set; }
+    
+    public void HideMouseCursor() => Dispatcher.UIThread.Post(() => ShowCursor = false);
+
+    public void ShowMouseCursor() => Dispatcher.UIThread.Post(() => ShowCursor = true);
 
     [ObservableProperty]
     [NotifyCanExecuteChangedFor(nameof(ShowPerformanceCommand))]
@@ -199,7 +274,6 @@ public sealed partial class MainWindowViewModel : ViewModelBase, IGui, IDisposab
         if(!string.IsNullOrWhiteSpace(filePath) &&
             File.Exists(filePath)) {
             await RestartEmulatorWithNewProgram(filePath);
-
         }
         else if (Application.Current?.ApplicationLifetime is IClassicDesktopStyleApplicationLifetime desktop &&
             desktop.MainWindow is not null &&
@@ -299,28 +373,26 @@ public sealed partial class MainWindowViewModel : ViewModelBase, IGui, IDisposab
     [RelayCommand]
     public void ResetTimeMultiplier() => TimeMultiplier = Configuration.TimeMultiplier;
 
-    public void UpdateScreen() {
-        if (_disposed || _isSettingResolution || _isMainWindowClosing) {
-            return;
-        }
-        VideoBuffer?.Draw();
-    }
-
-    public int Height { get; private set; }
+    public void UpdateScreen() => Draw();
 
     public double MouseX { get; set; }
-
     public double MouseY { get; set; }
-    public int Width { get; private set; }
 
     public bool IsLeftButtonClicked { get; private set; }
 
     public bool IsRightButtonClicked { get; private set; }
-    
-    public void OnMainWindowOpened() {
+
+    public void OnMainWindowInitialized(Action invalidateImage) {
+        SetUIUpdateMethod(invalidateImage);
         if(RunEmulator()) {
             _closeAppOnEmulatorExit = true;
         }
+    }
+
+    public void OnMainWindowShown(double imageWidth, double imageHeight) {
+        Width = (int)imageWidth;
+        Height = (int)imageHeight;
+        Bitmap = new WriteableBitmap(new PixelSize(Width, Height), new Vector(96, 96), PixelFormat.Bgra8888, AlphaFormat.Opaque);
     }
 
     private void AddOrReplaceMostRecentlyUsed(string filePath) {
@@ -376,19 +448,15 @@ public sealed partial class MainWindowViewModel : ViewModelBase, IGui, IDisposab
 
     public void SetResolution(int width, int height) => Dispatcher.UIThread.Post(() => {
         _isSettingResolution = true;
-        DisposeVideoBuffer();
-        Width = width;
-        Height = height;
-        IVideoBufferViewModel videoBuffer = new VideoBufferViewModel(_videoCard, scale: 1, width, height);
-        Dispatcher.UIThread.Post(() => {
-            VideoBuffer?.Dispose();
-            VideoBuffer = videoBuffer;
+        Scale = 1;
+        if (Width != width || Height != height) {
+            Width = width;
+            Height = height;
+            Bitmap?.Dispose();
+            Bitmap = new WriteableBitmap(new PixelSize(Width, Height), new Vector(96, 96), PixelFormat.Bgra8888, AlphaFormat.Opaque);
         }
-        );
         _isSettingResolution = false;
     }, DispatcherPriority.MaxValue);
-
-    private void DisposeVideoBuffer() => VideoBuffer?.Dispose();
 
     public void Dispose() {
         // Do not change this code. Put cleanup code in 'Dispose(bool disposing)' method
@@ -399,6 +467,14 @@ public sealed partial class MainWindowViewModel : ViewModelBase, IGui, IDisposab
     private void Dispose(bool disposing) {
         if (!_disposed) {
             if (disposing) {
+                _exitDrawThread = true;
+                if (_drawThread?.IsAlive == true) {
+                    _drawThread.Join();
+                }
+                Dispatcher.UIThread.Post(() => {
+                    Bitmap?.Dispose();
+                    Cursor?.Dispose();
+                }, DispatcherPriority.MaxValue);
                 PlayCommand.Execute(null);
                 IsMachineRunning = false;
                 DisposeEmulator();
@@ -414,10 +490,7 @@ public sealed partial class MainWindowViewModel : ViewModelBase, IGui, IDisposab
         }
     }
 
-    private void DisposeEmulator() {
-        DisposeVideoBuffer();
-        _programExecutor?.Dispose();
-    }
+    private void DisposeEmulator() => _programExecutor?.Dispose();
 
     [ObservableProperty]
     private bool _isDialogVisible;
@@ -476,10 +549,7 @@ public sealed partial class MainWindowViewModel : ViewModelBase, IGui, IDisposab
         ShowEmulationErrorMessage(e);
     });
 
-    [ObservableProperty]
-    private bool _showVideo = true;
-
-    private IVideoCard _videoCard = null!;
+    private IVideoCard? _videoCard;
 
     private void MachineThread() {
         try {
@@ -513,7 +583,6 @@ public sealed partial class MainWindowViewModel : ViewModelBase, IGui, IDisposab
         _videoCard = _programExecutor.Machine.VgaCard;
         Dispatcher.UIThread.Post(() => IsMachineRunning = true);
         Dispatcher.UIThread.Post(() => StatusMessage = "Emulator started.");
-        Dispatcher.UIThread.Post(() => ShowVideo = true);
         _programExecutor.Run();
         if (_closeAppOnEmulatorExit &&
             Application.Current?.ApplicationLifetime is IClassicDesktopStyleApplicationLifetime {
