+++ src/Spice86/Views/MainWindow.axaml.cs
@@ -2,6 +2,7 @@ namespace Spice86.Views;
 
 using Avalonia.Controls;
 using Avalonia.Input;
+using Avalonia.Interactivity;
 
 using Spice86.ViewModels;
 
@@ -10,14 +11,23 @@ using System.ComponentModel;
 internal partial class MainWindow : Window {
     public MainWindow() {
         InitializeComponent();
-        Closing += MainWindow_Closing;
     }
 
-    private Image? _videoBufferImage;
+    private void InvalidateImage() {
+        Image.InvalidateVisual();
+    }
 
     protected override void OnOpened(EventArgs e) {
         base.OnOpened(e);
-        (DataContext as MainWindowViewModel)?.OnMainWindowOpened();
+        FocusOnVideoBuffer();
+        var mainVm = (MainWindowViewModel?)DataContext;
+        Image.PointerMoved -= (s, e) => mainVm?.OnMouseMoved(e, Image);
+        Image.PointerPressed -= (s, e) => mainVm?.OnMouseButtonDown(e, Image);
+        Image.PointerReleased -= (s, e) => mainVm?.OnMouseButtonUp(e, Image);
+        Image.PointerMoved += (s, e) => mainVm?.OnMouseMoved(e, Image);
+        Image.PointerPressed += (s, e) => mainVm?.OnMouseButtonDown(e, Image);
+        Image.PointerReleased += (s, e) => mainVm?.OnMouseButtonUp(e, Image);
+        mainVm?.OnMainWindowInitialized(this.InvalidateImage);
     }
 
     protected override void OnClosed(EventArgs e) {
@@ -25,19 +35,10 @@ internal partial class MainWindow : Window {
         base.OnClosed(e);
     }
 
-    public void SetPrimaryDisplayControl(Image image) {
-        if(_videoBufferImage != image) {
-            _videoBufferImage = image;
-        }
-        FocusOnVideoBuffer();
-    }
-
     private void FocusOnVideoBuffer() {
-        if (_videoBufferImage is not null) {
-            _videoBufferImage.IsEnabled = false;
-            _videoBufferImage.Focus();
-            _videoBufferImage.IsEnabled = true;
-        }
+        Image.IsEnabled = false;
+        Image.Focus();
+        Image.IsEnabled = true;
     }
 
     protected override void OnKeyUp(KeyEventArgs e) {
@@ -50,7 +51,10 @@ internal partial class MainWindow : Window {
         FocusOnVideoBuffer();
     }
 
-    public static event EventHandler<CancelEventArgs>? AppClosing;
+    protected override void OnClosing(WindowClosingEventArgs e) {
+        AppClosing?.Invoke(this, e);
+        base.OnClosing(e);
+    }
 
-    private void MainWindow_Closing(object? sender, CancelEventArgs e) => AppClosing?.Invoke(sender, e);
+    public static event EventHandler<CancelEventArgs>? AppClosing;
 }
